Add native demos
